backend/requirements.txt
Flask==2.3.2 
Flask-SQLAlchemy==3.0.5 
Flask-CORS==4.0.0 
Flask-JWT-Extended==4.5.2 
Werkzeug==3.0.1 
pytest==7.4.0 
gunicorn==21.2.0